Match overflow calculation in StringBuilder::AddSubstring with the one in FakeOFStream::operator<<
StringBuilder::AddSubstring assumes that the substring being added is smaller than the buffer. While technically there is space for data exactly the size of buffer, if I understand StringBuilder::Finalize correctly it will use that last spot for a terminating null-byte. And while Finalize() is not used anywhere in the current code base, it is called by the StringBuilder destructor. This fix leaves all of that intact and just moves the limit in FakeOFStream a bit.